Moved parsing functions to a mixin
